requirements.txt
robotframework
robotframework-browser
robotframework-requests
robotframework-tidy
robotframework-otp
pyotp
python-dotenv